fixup! client/local,ipn/localapi: add /localapi/v0/routecheck endpoint
@bradfitz pointed out that I was holding localapi.Register upside down.
